move create post to a separate page
Although I do want this web app to be an SPA,
I don't want to get lost in the weeds of client-
side JavaScript nor do I want to shoot myself in
the foot by trying to use SPA patterns in the server
(I already saw how that caused me to not give the
renderIndex route a single responsibility, and it
was causing me to violate the open-closed principle).
So, I'm creating this a normal website for now.
I'll migrate it to an SPA much much later.