Attempt at improving dependency management.
I was comfortable serving my client-side JS
dependencies as static files hosted on the server.
This worked for Leaflet, as it was a UMD. But then
I wanted to use Redux to improve my client-side
state management, and I found that the latest
version of Redux uses ES modules, so I'd have to
change my approach.

Here, I try to leverage Deno, installing the
dependencies with it and using its experimental
bundler. And, although I'm not getting any errors,
the map is broken. My guess is that the Deno bundler
is not including all of the Leaflet package.