move search suggestion generation to the server
When testing the website last night, I noticed
that as the number of items went up, loading times
slowed dramatically. This was since the index page
would be served with all the items in the database.
I'm changing that so that the index page is only
served with one item, giving the two-fold benefit
of faster loading times and the ability to bookmark
map views of certain items. I thought that the
search suggestion feature was a good place to start
implementing this change.