limbo
I was mid-way through making the map navigation
buttons work again. Starting from the server-side,
I (with help from ChatGPT) wrote the queries to
select the primary keys of the item preceding and
the item following a given item, and I passed the
object containing these into the template. But I
haven't been able to test this yet.

Looking at the client side, I still wanted the
seamless SPA experience when navigating different
items across the map, so I'd need to fetch the
item data when a navigation button is clicked and
render it. I wanted to use EJS for the rendering
as I was already using it in the server side.
Unfortunately, it's been impossible for me to
import it, so I'm in limbo now. I'll just post an
issue to the EJS Github (like I ultimately did for
the leaflet issue), and see.